@@ -1,114 +1,114 @@
-import React from 'react';
-import './Footer.css';
-import { Button } from './Button';
-import { Link } from 'react-router-dom';
+import React from "react";
+import "./Footer.css";
+import { Button } from "./Button";
+import { Link } from "react-router-dom";
 
 function Footer() {
   return (
-    <div className='footer-container'>
-      <section className='footer-subscription'>
-        <p className='footer-subscription-heading'>
+    <div className="footer-container">
+      <section className="footer-subscription">
+        <p className="footer-subscription-heading">
           Join the Adventure newsletter to receive our best vacation deals
         </p>
-        <p className='footer-subscription-text'>
+        <p className="footer-subscription-text">
           You can unsubscribe at any time.
         </p>
-        <div className='input-areas'>
+        <div className="input-areas">
           <form>
             <input
-              className='footer-input'
-              name='email'
-              type='email'
-              placeholder='Your Email'
+              className="footer-input"
+              name="email"
+              type="email"
+              placeholder="Your Email"
             />
-            <Button buttonStyle='btn--outline'>Subscribe</Button>
+            <Button buttonStyle="btn--outline">Subscribe</Button>
           </form>
         </div>
       </section>
-      <div class='footer-links'>
-        <div className='footer-link-wrapper'>
-          <div class='footer-link-items'>
+      <div class="footer-links">
+        <div className="footer-link-wrapper">
+          <div class="footer-link-items">
             <h2>About Us</h2>
-            <Link to='/sign-up'>How it works</Link>
-            <Link to='/'>Testimonials</Link>
-            <Link to='/'>Careers</Link>
-            <Link to='/'>Investors</Link>
-            <Link to='/'>Terms of Service</Link>
+            <Link to="/sign-up">How it works</Link>
+            <Link to="/">Testimonials</Link>
+            <Link to="/">Careers</Link>
+            <Link to="/">Investors</Link>
+            <Link to="/">Terms of Service</Link>
           </div>
-          <div class='footer-link-items'>
+          <div className="footer-link-items">
             <h2>Contact Us</h2>
-            <Link to='/'>Contact</Link>
-            <Link to='/'>Support</Link>
-            <Link to='/'>Destinations</Link>
-            <Link to='/'>Sponsorships</Link>
+            <Link to="/">Contact</Link>
+            <Link to="/">Support</Link>
+            <Link to="/">Destinations</Link>
+            <Link to="/">Sponsorships</Link>
           </div>
         </div>
-        <div className='footer-link-wrapper'>
-          <div class='footer-link-items'>
+        <div className="footer-link-wrapper">
+          <div className="footer-link-items">
             <h2>Videos</h2>
-            <Link to='/'>Submit Video</Link>
-            <Link to='/'>Ambassadors</Link>
-            <Link to='/'>Agency</Link>
-            <Link to='/'>Influencer</Link>
+            <Link to="/">Submit Video</Link>
+            <Link to="/">Ambassadors</Link>
+            <Link to="/">Agency</Link>
+            <Link to="/">Influencer</Link>
           </div>
-          <div class='footer-link-items'>
+          <div className="footer-link-items">
             <h2>Social Media</h2>
-            <Link to='/'>Instagram</Link>
-            <Link to='/'>Facebook</Link>
-            <Link to='/'>Youtube</Link>
-            <Link to='/'>Twitter</Link>
+            <Link to="/">Instagram</Link>
+            <Link to="/">Facebook</Link>
+            <Link to="/">Youtube</Link>
+            <Link to="/">Twitter</Link>
           </div>
         </div>
       </div>
-      <section class='social-media'>
-        <div class='social-media-wrap'>
-          <div class='footer-logo'>
-            <Link to='/' className='social-logo'>
-            Manir & Tania 
-              <i class='fab fa-typo3' />
+      <section className="social-media">
+        <div className="social-media-wrap">
+          <div className="footer-logo">
+            <Link to="/" className="social-logo">
+              Manir & Tania
+              <i className="fab fa-typo3" />
             </Link>
           </div>
-          <small class='website-rights'>Manir & Tania © 2020</small>
-          <div class='social-icons'>
+          <small className="website-rights">Manir & Tania © 2020</small>
+          <div className="social-icons">
             <Link
-              class='social-icon-link facebook'
-              to='/'
-              target='_blank'
-              aria-label='Facebook'
+              className="social-icon-link facebook"
+              to="/"
+              target="_blank"
+              aria-label="Facebook"
             >
-              <i class='fab fa-facebook-f' />
+              <i className="fab fa-facebook-f" />
             </Link>
             <Link
-              class='social-icon-link instagram'
-              to='/'
-              target='_blank'
-              aria-label='Instagram'
+              className="social-icon-link instagram"
+              to="/"
+              target="_blank"
+              aria-label="Instagram"
             >
-              <i class='fab fa-instagram' />
+              <i className="fab fa-instagram" />
             </Link>
             <Link
-              class='social-icon-link youtube'
-              to='/'
-              target='_blank'
-              aria-label='Youtube'
+              className="social-icon-link youtube"
+              to="/"
+              target="_blank"
+              aria-label="Youtube"
             >
-              <i class='fab fa-youtube' />
+              <i className="fab fa-youtube" />
             </Link>
             <Link
-              class='social-icon-link twitter'
-              to='/'
-              target='_blank'
-              aria-label='Twitter'
+              className="social-icon-link twitter"
+              to="/"
+              target="_blank"
+              aria-label="Twitter"
             >
-              <i class='fab fa-twitter' />
+              <i className="fab fa-twitter" />
             </Link>
             <Link
-              class='social-icon-link twitter'
-              to='/'
-              target='_blank'
-              aria-label='LinkedIn'
+              className="social-icon-link twitter"
+              to="/"
+              target="_blank"
+              aria-label="LinkedIn"
             >
-              <i class='fab fa-linkedin' />
+              <i className="fab fa-linkedin" />
             </Link>
           </div>
         </div>
